% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearbyCities.R
\name{findNearbyCities}
\alias{findNearbyCities}
\title{Find nearby cities within a certain distance from a location specified by longtitude and latitude}
\usage{
findNearbyCities(username, coord, distance, maxRows)
}
\arguments{
\item{username}{username to access geonames.org api}

\item{coord}{a list of character to specify the location in longtitude and latitude}

\item{distance}{a character to specify distance in km}

\item{maxRows}{a character to specify the max number of nearby cities that need to be returned}
}
\value{
a vector of cities
}
\description{
Find nearby cities within a certain distance from a location specified by longtitude and latitude.
}
\examples{
findNearbyCities('nan_stat290', longlat('San Francisco, CA'), '100', '5')
}
