% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlat.R
\name{longlat}
\alias{longlat}
\title{Find longtitude and latitude coordinates of a given address}
\usage{
longlat(addr)
}
\arguments{
\item{addr}{a character vector of an address}
}
\value{
coord longitude and latitude in character format. It will stop if it cannot find the address
}
\description{
find longtitude and latitude coordinates of a given address
}
\examples{
longlat("San Francisco, US")
}
