% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pws.R
\docType{class}
\name{Pws}
\alias{Pws}
\title{An R6 Class to represent one or more Personal Weather Stations.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Pws
}
\description{
An R6 Class to represent one or more Personal Weather Stations
}
\section{Fields}{

\describe{
\item{\code{id}}{A character vector containing 1 or more PWS ids.  REQUIRED.}

\item{\code{lat}}{A numeric vector containing lat coordiantes for PWS ids.  REQUIRED.}

\item{\code{lon}}{A numeric vector containing lat coordiantes for PWS ids.  REQUIRED.}

\item{\code{neighborhood}}{A character vector containing the neighborhoods of the PWSs (1 per PWS)}

\item{\code{city}}{A character vector containing the cities of the PWSs (1 per PWS)}

\item{\code{state}}{A character vector containing state codes of the PWSs (1 per PWS)}

\item{\code{country}}{A character vector containing country codes of the PWSs (1 per PWS)}

\item{\code{distance_km}}{An integer vector of the distance in km for each PWS from city.  REQUIRED, but has a 0 default.}

\item{\code{distance_mi}}{An integer vector of the distance in miles for each PWS from city}
}}

\section{Methods}{

\describe{
  \item{\code{getId(pwsId)}}{This method uses \code{pwsId} to determine if to get some or all (if NA) of the ids in the class instance}
  \item{\code{getLat(pwsId)}}{This method uses \code{pwsId} to determine if to get some or all (if NA) of the latitudes in the class instance}
  \item{\code{getLon(pwsId)}}{This method uses \code{pwsId} to determine if to get some or all (if NA) of the longitudes in the class instance}
  \item{\code{getCity(pwsId)}}{This method uses \code{pwsId} to determine if to get some or all (if NA) of the cities in the class instance}
  \item{\code{getNeighborhood(pwsId)}}{This method uses \code{pwsId} to determine if to get some or all (if NA) of the neighborhoods in the class instance}
  \item{\code{getState(pwsId)}}{This method uses \code{pwsId} to determine if to get some or all (if NA) of the states in the class instance}
  \item{\code{getCountry(pwsId)}}{This method uses \code{pwsId} to determine if to get some or all (if NA) of the countries in the class instance}
  \item{\code{getDistance_km(pwsId)}}{This method uses \code{pwsId} to determine if to get some or all (if NA) of the distances in kms in the class instance}
  \item{\code{getDistance_mi(pwsId)}}{This method uses \code{pwsId} to determine if to get some or all (if NA) of the distances in miles in the class instance}
  \item{\code{getTibble(pwsId,pwsCity,pwsState,pwsDistanceLte)}}{This method uses \code{pwsId} as  priority 1 by itself/only, \code{state} priority 2 (assumed in US) and/or \code{country} priority 3, then \code{city}, and \code{pwsDistanceLte} in km to return a tibble containing pws data}
  \item{\code{print()}}{This method prints the class in the manner or a tibble}
  \item{\code{plot(plottype)}}{This method \code{plottype} to determine whether p/pins or t/text (ids) to produce a plot of weather stations on a map using RgoogleMaps package}
}
}

\examples{
data(pwsData)
testr6 <- Pws$new(
  id               = pwsData$id
  ,lat             = pwsData$lat
  ,lon             = pwsData$lon
  ,neighborhood    = pwsData$neighborhood
  ,city            = pwsData$city
  ,state           = pwsData$state
  ,country         = pwsData$country
  ,distance_km     = pwsData$distance_km
  ,distance_mi     = pwsData$distance_mi
)
testr6$print()
testr6$plot()
}
\keyword{datasets}
