% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryData.R
\name{queryData}
\alias{queryData}
\title{Query weather data for individual PWS}
\usage{
queryData(myKey, id, qTime)
}
\arguments{
\item{myKey}{key for accessing Wunderground API}

\item{id}{character pws ID}

\item{qTime}{a specific date in history for query}
}
\value{
a tibble containing the data for that specified pws from the specified date
}
\description{
Query weather data for individual PWS
}
\examples{
queryData("407e6151ab7de146", "KCASANFR699", Sys.Date()-1)
}
