% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHistory.R
\name{queryHistory}
\alias{queryHistory}
\title{Query weather history data from the specified PWS within the specified date range}
\usage{
queryHistory(myKey, nearbyStations, startDate, endDate)
}
\arguments{
\item{myKey}{key for accessing Wunderground API}

\item{nearbyStations}{a tibble describing PWS}

\item{startDate}{a starting date}

\item{endDate}{an end date}
}
\value{
a tibble containing weather data for all the pws within the specified date range
}
\description{
Query weather history data from the specified PWS within the specified date range
}
\examples{
data(pwsData)
queryHistory("407e6151ab7de146", pwsData, Sys.Date()-1, Sys.Date())
}
