% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeData.R
\name{summarizeData}
\alias{summarizeData}
\title{Aggregate weather data from pws within certain distance and specified date range}
\usage{
summarizeData(weatherData, distance, startDate, endDate, param, unit, calType)
}
\arguments{
\item{weatherData}{a tibble containing some history weather data for some pws}

\item{distance}{a numeric to specify distance in km}

\item{startDate}{a starting date}

\item{endDate}{an end date}

\item{param}{a weather parameter of interest}

\item{unit}{either 'metric' or 'english'}

\item{calType}{the statistics of interest: mean, max, min, range}
}
\value{
a tibble containing the results
}
\description{
Aggregate weather data from pws within certain distance and specified date range
}
\examples{
data(weatherData)
summarizeData(weatherData, 10, '2017-03-18', '2017-03-19', 'temperature', 'metric', 'mean')
}
