% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weatherData}
\alias{weatherData}
\title{a data set describing weather data}
\format{An object of class \code{data.frame} with 3758 rows and 32 columns.}
\usage{
weatherData
}
\description{
The \code{weatherData} dataset for plotWeather and summarizeData function
}
\keyword{data}
