% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherProject.R
\docType{package}
\name{weatherProject}
\alias{weatherProject}
\alias{weatherProject-package}
\title{A package for Wunderground weather project}
\description{
A collection of simple functions for illustration. See examples
for functions \code{link{findNearbyCities}} and \code{link{queryData}}.
}
\seealso{
\code{\link{findNearbyCities}}

\code{link{findPWS}}

\code{link{longlat}}

\code{link{plotWeather}}

\code{link{queryData}}

\code{link{queryHistory}}

\code{link{summarizeData}}
}
