% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeather.R
\name{plotWeather}
\alias{plotWeather}
\title{Plot weather data for visualization}
\usage{
plotWeather(wtbl, plottype = "m", wvar = "tempi", aggtype = "mean",
  startDate = NA, endDate = NA)
}
\arguments{
\item{wtbl}{tibble or data frame of weather data must contain id,lat,lon & all types of wvar}

\item{plottype}{specify plot type. m for mapping; g for graphs.}

\item{wvar}{weather var - tempm,tempi,dewptm,dewpti,pressurem,pressurei,hum, m=metric;i=english}

\item{aggtype}{aggregation type to roll-up the wvar - can be mean,max,min,range}

\item{startDate}{start date - will be applied to the date portion of a variable called date_time}

\item{endDate}{end date - will be applied to the date portion of a variable called date_time}
}
\description{
Plot weather data for visualization
}
\examples{
data(weatherData)
plotWeather(weatherData)
plotWeather(wtbl=weatherData,wvar="tempi",aggtype="max", plottype = "g")
plotWeather(wtbl=weatherData,aggtype="range",startDate = '2017-03-18',endDate ='2017-03-19')
}
