% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeather.R
\name{plotWeather}
\alias{plotWeather}
\title{Plot weather data for visualization}
\usage{
plotWeather(wtbl, plottype = "m", wvar = "tempi", aggtype = "mean",
  startDate = NA, endDate = NA)
}
\arguments{
\item{wtbl}{a tibble of weather data}

\item{plottype}{specify plot type}

\item{wvar}{weather parameter of interest}

\item{aggtype}{aggregate type}

\item{startDate}{start date}

\item{endDate}{end date}
}
\description{
Plot weather data for visualization
}
\examples{
data(weatherData)
plotWeather(weatherData, plottype = 'm')
}
